+++ src/trading_pair.rs
@@ -1,0 +1,7 @@
+use strum_macros::{Display, EnumIter};
+
+#[derive(Clone, Debug, Display, EnumIter)]
+pub enum ETradingPair {
+    BtcUsdc,
+    SolUsdc
+}