+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2024-08-22
+
+### Added
+
+- Implemented order book management for `Bybit` exchange via WebSocket
+- Implemented automatic reconnection mechanism in case of WebSocket errors
+- Added ability to select trading pair at program startup
+
 ## [0.0.2] - 2024-08-20
 
 ### Added